Remove all non-twitter-handle characters
Different twitter clients add different invisible characters 
to the ends of the twitter handle when you copy-paste, 
causing annoying errors that are difficult to avoid 
manually. This change just skips all characters that are
not alphanumeric or underscores, which are the only
characters twitter allows in a handle. The @ symbol is
optional but is not required so we skip that too.